add starrocks writer
