--- doriswriter/src/main/assembly/package.xml
@@ -1,35 +1,0 @@
-<assembly
-	xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0 http://maven.apache.org/xsd/assembly-1.1.0.xsd">
-	<id></id>
-	<formats>
-		<format>dir</format>
-	</formats>
-	<includeBaseDirectory>false</includeBaseDirectory>
-	<fileSets>
-		<fileSet>
-			<directory>src/main/resources</directory>
-			<includes>
-				<include>plugin.json</include>
-				<include>plugin_job_template.json</include>
-			</includes>
-			<outputDirectory>plugin/writer/doriswriter</outputDirectory>
-		</fileSet>
-		<fileSet>
-			<directory>target/</directory>
-			<includes>
-				<include>doriswriter-0.0.1-SNAPSHOT.jar</include>
-			</includes>
-			<outputDirectory>plugin/writer/doriswriter</outputDirectory>
-		</fileSet>
-	</fileSets>
-
-	<dependencySets>
-		<dependencySet>
-			<useProjectArtifact>false</useProjectArtifact>
-			<outputDirectory>plugin/writer/doriswriter/libs</outputDirectory>
-			<scope>runtime</scope>
-		</dependencySet>
-	</dependencySets>
-</assembly>
