--- doriswriter/src/main/resources/plugin.json
@@ -1,6 +1,0 @@
-{
-    "name": "doriswriter",
-    "class": "com.dorisdb.connector.datax.plugin.writer.doriswriter.DorisWriter",
-    "description": "useScene: prod. mechanism: DorisStreamLoad. warn: The more you know about the database, the less problems you encounter.",
-    "developer": "dorisdb"
-}
