--- doriswriter/src/main/resources/plugin_job_template.json
@@ -1,14 +1,0 @@
-{
-    "name": "doriswriter",
-    "parameter": {
-        "username": "",
-        "password": "",
-        "database": "",
-        "table": "",
-        "column": [],
-        "preSql": [],
-        "postSql": [], 
-        "jdbcUrl": "",
-        "loadUrl": []
-    }
-}
