+++ package.xml
@@ -190,7 +190,7 @@
             <outputDirectory>datax</outputDirectory>
         </fileSet>
         <fileSet>
-            <directory>doriswriter/target/datax/</directory>
+            <directory>starrockswriter/target/datax/</directory>
             <includes>
                 <include>**/*.*</include>
             </includes>
