+++ starrockswriter/src/main/java/com/starrocks/connector/datax/plugin/writer/doriswriter/row/StarRocksBaseSerializer.java
@@ -1,0 +1,18 @@
+package com.starrocks.connector.datax.plugin.writer.starrockswriter.row;
+
+import com.alibaba.datax.common.element.Column;
+import com.alibaba.datax.common.element.Column.Type;
+
+public class StarRocksBaseSerializer {
+
+    protected String fieldConvertion(Column col) {
+        if (null == col.getRawData()) {
+            return null;
+        }
+        if (Type.BOOL == col.getType()) {
+            return String.valueOf(col.asLong());
+        }
+        return col.asString();
+    }
+
+}
