+++ starrockswriter/src/main/java/com/starrocks/connector/datax/plugin/writer/doriswriter/row/StarRocksISerializer.java
@@ -1,0 +1,11 @@
+package com.starrocks.connector.datax.plugin.writer.starrockswriter.row;
+
+import java.io.Serializable;
+
+import com.alibaba.datax.common.element.Record;
+
+public interface StarRocksISerializer extends Serializable {
+
+    String serialize(Record row);
+    
+}
